refactor(platform): refactor api validatiion logic
