+++ pkg/platform/provider/baremetal/cluster/util.go
@@ -22,17 +22,13 @@ import (
 	"fmt"
 	"math"
 	"net"
-	"reflect"
 
-	mapset "github.com/deckarep/golang-set"
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/util/sets"
 	utilsnet "k8s.io/utils/net"
-	"tkestack.io/tke/api/platform"
 	platformv1 "tkestack.io/tke/api/platform/v1"
 	"tkestack.io/tke/pkg/platform/provider/baremetal/constants"
 	"tkestack.io/tke/pkg/util/ipallocator"
-	"tkestack.io/tke/pkg/util/log"
 )
 
 func GetNodeCIDRMaskSize(clusterCIDR string, maxNodePodNum int32) (int32, error) {
@@ -141,51 +137,4 @@ func CalcNodeCidrSize(podSubnet string) (int32, bool) {
 		}
 	}
 	return int32(maskSize), isIPv6
-}
-
-func PrepareClusterScale(cluster *platform.Cluster, oldCluster *platform.Cluster) ([]platform.ClusterMachine, error) {
-	allMachines, scalingMachines := []platform.ClusterMachine{}, []platform.ClusterMachine{}
-
-	oIPs := mapset.NewSet()
-	for _, machine := range oldCluster.Spec.Machines {
-		oIPs.Add(machine.IP)
-		allMachines = append(allMachines, machine)
-	}
-	IPs := mapset.NewSet()
-	for _, machine := range cluster.Spec.Machines {
-		IPs.Add(machine.IP)
-		allMachines = append(allMachines, machine)
-	}
-	// nothing to do since ips not changed
-	if reflect.DeepEqual(oIPs, IPs) {
-		return scalingMachines, nil
-	}
-	// machine in oldCluster but not in cluster
-	diff1 := oIPs.Difference(IPs)
-	// machine in cluster but not in oldCluster
-	diff2 := IPs.Difference(oIPs)
-	// scaling machine ips
-	diff := mapset.NewSet()
-	log.Errorf("PrepareClusterScale called: diff1 -> %v, diff2 -> %v", diff1.ToSlice(), diff2.ToSlice())
-	if diff1.Cardinality() > 0 && diff2.Cardinality() > 0 {
-		return scalingMachines, errors.Errorf("scale up and down master in parallel is not allowed: %v, %v", diff1.ToSlice(), diff2.ToSlice())
-	}
-	if diff1.Cardinality() > 0 {
-		if diff1.Contains(oldCluster.Spec.Machines[0].IP) {
-			return scalingMachines, errors.Errorf("master[0] can't scale down: %v", oldCluster.Spec.Machines[0].IP)
-		}
-		diff = diff1
-	}
-	if diff2.Cardinality() > 0 {
-		diff = diff2
-	}
-	for _, m := range diff.ToSlice() {
-		for _, machine := range allMachines {
-			if m == machine.IP {
-				scalingMachines = append(scalingMachines, machine)
-				break
-			}
-		}
-	}
-	return scalingMachines, nil
 }
