+++ pkg/platform/provider/baremetal/machine/provider.go
@@ -19,6 +19,10 @@
 package machine
 
 import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/client-go/tools/clientcmd"
 	platformv1client "tkestack.io/tke/api/client/clientset/versioned/typed/platform/v1"
@@ -123,5 +127,17 @@ func NewProvider() (*Provider, error) {
 var _ machineprovider.Provider = &Provider{}
 
 func (p *Provider) Validate(machine *platform.Machine) field.ErrorList {
-	return validation.ValidateMachine(machine)
+	allErrs := field.ErrorList{}
+	fldPath := field.NewPath("spec")
+
+	cluster, err := p.platformClient.Clusters().Get(context.TODO(), machine.Spec.ClusterName, metav1.GetOptions{})
+	if err != nil {
+		if errors.IsNotFound(err) {
+			allErrs = append(allErrs, field.NotFound(fldPath.Child("ClusterName"), machine.Spec.ClusterName))
+		} else {
+			allErrs = append(allErrs, field.InternalError(fldPath, err))
+		}
+	}
+
+	return append(allErrs, validation.ValidateMachine(machine, cluster, p.platformClient)...)
 }
